requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
asttokens==3.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
blis==1.3.3
build==1.3.0
cachetools==6.2.2
catalogue==2.0.10
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
cformers==0.0.4
charset-normalizer==3.4.4
chroma-hnswlib==0.7.3
chromadb==1.3.5
click==8.3.1
click-default-group==1.2.4
cloudpathlib==0.23.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
confection==0.1.5
cryptography==46.0.3
cssselect2==0.8.0
cymem==2.0.13
dataclasses-json==0.6.7
datasets==4.4.1
debugpy==1.8.18
decorator==5.2.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
docstring_parser==0.17.0
durationpy==0.10
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.122.1
filelock==3.20.0
flatbuffers==25.9.23
freetype-py==2.5.1
frozenlist==1.8.0
fsspec==2025.10.0
google-auth==2.43.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
identify==2.6.15
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.13.0
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kubernetes==34.1.0
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-groq==0.3.8
langchain-huggingface==0.3.1
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.2.35
langgraph-checkpoint==2.1.2
langsmith==0.4.56
lxml==6.0.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals==2.12.0
nest-asyncio==1.6.0
networkx==3.6.1
nodeenv==1.9.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.8.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-asgi==0.59b0
opentelemetry-instrumentation-fastapi==0.59b0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
opentelemetry-util-http==0.59b0
orjson==3.11.5
ormsgpack==1.12.0
oscrypto==1.3.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
passlib==1.7.4
pillow==12.0.0
platformdirs==4.5.1
plotly==6.5.0
pluggy==1.6.0
posthog==5.4.0
pre_commit==4.5.0
preshed==3.0.12
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pulsar-client==3.8.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pypdf==6.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
ragas==0.4.0
RapidFuzz==3.14.3
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.5
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rlPyCairo==0.4.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scikit-network==0.33.5
scipy==1.16.3
sentence-transformers==5.1.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
sqlite-fts4==1.0.3
sqlite-utils==3.39
srsly==2.5.2
stack-data==0.6.3
starlette==0.50.0
svglib==1.6.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
thinc==8.3.10
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.5.1
tokenizers==0.22.1
torch==2.9.1
tornado==6.5.3
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
ty==0.0.1a32
typer==0.20.0
typer-slim==0.20.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==6.0.1
urllib3==2.3.0
uuid_utils==0.12.0
uvicorn==0.38.0
virtualenv==20.35.4
wasabi==1.1.3
watchfiles==1.1.1
wcwidth==0.2.14
weasel==0.4.3
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wget==3.2
wrapt==1.17.3
xhtml2pdf==0.2.17
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
